Pages/SistemaV2/Backend/requirements.txt
# ============================================================
# REGISTRA PERÚ - SistemaV2
# Requerimientos Backend (Optimizado para Python 3.14)
# ============================================================

# 🔧 Configuración
python-dotenv==1.0.1
requests==2.32.3

# 🤖 Telegram Bot
Telethon==1.34.0

# 🌐 API principal
fastapi==0.115.0
uvicorn==0.30.6

# 🧩 Manejo de archivos y multipart
aiofiles==24.1.0
python-multipart==0.0.9

# 🖼️ Imágenes (compatible con Python 3.14)
Pillow==12.0.0

# 🔄 Utilidades adicionales
tqdm==4.66.5
pydantic==2.12.2
typing-extensions==4.15.0
